Load expanded !info if there is only one result in !search
