+++ backend/src/plugins/Utility/types.ts
@@ -24,6 +24,7 @@ export const ConfigSchema = t.type({
   can_jumbo: t.boolean,
   jumbo_size: t.Integer,
   can_avatar: t.boolean,
+  info_on_single_result: t.boolean,
 });
 export type TConfigSchema = t.TypeOf<typeof ConfigSchema>;
